refactor(pages): extract page-level UI and logic into reusable sections and hooks to keep page files lean
